New discriminants may be part of the current description.
Commented the lines when the current properties were introduced in the list of
ignored features. For instance, asking for the type is useful, even if in the
query it already exists.

The query may come for inference: ex. give me the box. The object TRASHBIN is
type trashbin, but box is a superclass for trashbin. Thus TRASHBIN will be both
types, box and trashbin. The query refers to the type box, but not to the type
trashbin.

To quickly solve this problem I'm using the getDirectClass function. However,
a first step to really search for correct values will be added in the following commit.
This is just to solve the problem for the ROMAN demo.

DiscriminateForAgent (different from the robot) is not working. So, for now
we only look at discriminants in the robot's model (although the objects have been
obtained using the description in the agent's model).