Associate an 'InterrogativeState' to question; delay answer to human
This commit changes the way questions are answered: instead of directly verbalizing the answer to
the interactor, a new interrogative 'state of mind' is created, associated to the human and stored
in the knowledge base, along with the answer the dialog module has computed.

The robot controller is then in charge of monitoring this 'InterrogativeState' and react
appropriately, by telling the answer or maybe triggering other behaviours (for instance, start
lokking for an object if it is not found).

As a side-effect of the commit, the dialogs module won't say anymore 'Ok' or 'Alright' after
an order. These short sentences are now available in the knowledge base as 'sit_id verbalisesTo <text>'