Correct semantics of indirect objects in actions
Indirect objects often do not refer to the action directly (ie, no direct thematic
role) but indirectly.

For instance: 'look on the table'
'on the table' do not qualify the 'look' action, but the zone which is looked.

This commit improves this situation: when no thematic role is explicitely found,
we say the action 'involves' a new entity which is qualified by the indirect cmplt.

For instance: 'look on the table' becomes:
>> vwTYU rdf:type Look
>> vwTYU performedBy myself
>> vwTYU rdfs:label "Look action #vwTYU"
>> TOTO desires vwTYU
>> vwTYU involves vXgtb
>> vXgtb isOn UNKNOWN_CONCEPT_IfzXw